feat(file): add context manager to File (#998)
* Add Context Manager

Add Context Manager to `nextcord.File`

* feat: add context manager for `File`

feat: add context manager for `File`

Add a context manager to `nextcord.File`. Also change `force_close` to be optional default to None so that we can set `force_close` to True when using context manager if `False` was not provided to init.

* feat: add context manager for `File`

feat: add context manager for `File`

Add a context manager to `nextcord.File`. Also change `force_close` to be optional default to None so that we can set `force_close` to True when using context manager if `False` was not provided to init.

* Update docstrings

Co-authored-by: Emre Terzioglu <50607143+EmreTech@users.noreply.github.com>
Co-authored-by: Oliver Wilkes <oliverwilkes2006@icloud.com>

* Update nextcord/file.py

Co-authored-by: Emre Terzioglu <50607143+EmreTech@users.noreply.github.com>

---------

Co-authored-by: Emre Terzioglu <50607143+EmreTech@users.noreply.github.com>
Co-authored-by: Oliver Wilkes <oliverwilkes2006@icloud.com>