Address readability issue using `super()`
If further discussion is made in favour of avoiding super,
this change could be reverted.
I still stand against using super as it is an unnecessary function call
when implicity statement could be used. super should be best reserved for
when there are multiple inheritences, and isn't necessary for this use case.

Changes to be committed:
	modified:   nextcord/cache.py